pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pkmai"
version = "0.1.0"
description = "Local AI orchestration for Personal Knowledge Management"
authors = [
    {name = "Mael Depréville", email = "maeldepreville@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "llama-cpp-python>=0.3.18",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "tqdm>=4.67.3",
    "typer>=0.24.1",
    "uvicorn>=0.42.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.8",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
pkmai = "pkmai.cli:app"
